Reorganised source files layout
